+++ app/lib/repo/impl/vehicles_repo_impl.dart
@@ -98,7 +98,7 @@ class VehiclesRepoImpl extends VehiclesRepo {
   @override
   Future<Result<List<Vehicle>>> loadVehiclesNearby(
     LatLng position, {
-    @required double radiusInMeters,
+    @required int radiusInMeters,
     int type,
   }) async {
     try {
